feat: add api versioning, restructure urls and views of the existing apps as v1, add versioning config(settings)
