+++ core/urls.py
@@ -3,7 +3,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("api/", include("systems.urls")),
-    path("api/measurements/", include("measurements.urls")),
     path("", include("auth_app.urls")),
+    path("api/v1/systems/", include("systems.api.v1.urls")),
+    path("api/v1/measurements/", include("measurements.api.v1.urls")),
 ]
