+++ measurements/api/v1/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from measurements.views import MeasurementListCreate
+from measurements.api.v1.views import MeasurementListCreate
 
 urlpatterns = [
     path("", MeasurementListCreate.as_view(), name="measurement-list-create"),
